writer/requirements.txt
Flask
beanstalkc3
redis
